requirements.txt
torch
transformers
Pillow
accelerate